# apps/fastapi/openapi.json
{
  "components": {
    "schemas": {
      "BatchScoreRequest": {
        "properties": {
          "points": {
            "items": {
              "$ref": "#/components/schemas/Point"
            },
            "maxItems": 5000,
            "minItems": 1,
            "title": "Points",
            "type": "array"
          },
          "preset": {
            "default": "warehouse",
            "enum": [
              "warehouse",
              "retail",
              "ev"
            ],
            "title": "Preset",
            "type": "string"
          },
          "weights": {
            "anyOf": [
              {
                "additionalProperties": {
                  "minimum": 0.0,
                  "type": "number"
                },
                "propertyNames": {
                  "enum": [
                    "demographics",
                    "transport",
                    "poi",
                    "zoning",
                    "flood",
                    "aqi"
                  ]
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weights"
          }
        },
        "required": [
          "points"
        ],
        "title": "BatchScoreRequest",
        "type": "object"
      },
      "BatchScoreResponse": {
        "properties": {
          "results": {
            "items": {
              "$ref": "#/components/schemas/ScoreResponse"
            },
            "title": "Results",
            "type": "array"
          }
        },
        "required": [
          "results"
        ],
        "title": "BatchScoreResponse",
        "type": "object"
      },
      "ConstraintResult": {
        "properties": {
          "actual": {
            "title": "Actual"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "label": {
            "default": "",
            "title": "Label",
            "type": "string"
          },
          "pass": {
            "title": "Pass",
            "type": "boolean"
          },
          "required": {
            "title": "Required"
          }
        },
        "required": [
          "id",
          "actual",
          "required",
          "pass"
        ],
        "title": "ConstraintResult",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HeatmapCell": {
        "description": "One heatmap cell: identity + weight-independent subscores only.\n\nNo composite score, no geometry \u2014 the frontend derives boundaries\nwith h3-js and recomputes the weighted composite locally.",
        "properties": {
          "eligible": {
            "title": "Eligible",
            "type": "boolean"
          },
          "h3Index": {
            "title": "H3Index",
            "type": "string"
          },
          "subscores": {
            "$ref": "#/components/schemas/Subscores"
          }
        },
        "required": [
          "h3Index",
          "eligible",
          "subscores"
        ],
        "title": "HeatmapCell",
        "type": "object"
      },
      "HeatmapResponse": {
        "properties": {
          "cells": {
            "items": {
              "$ref": "#/components/schemas/HeatmapCell"
            },
            "title": "Cells",
            "type": "array"
          },
          "datasetId": {
            "title": "Datasetid",
            "type": "string"
          },
          "h3Resolution": {
            "title": "H3Resolution",
            "type": "integer"
          },
          "preset": {
            "enum": [
              "warehouse",
              "retail",
              "ev"
            ],
            "title": "Preset",
            "type": "string"
          }
        },
        "required": [
          "datasetId",
          "h3Resolution",
          "preset",
          "cells"
        ],
        "title": "HeatmapResponse",
        "type": "object"
      },
      "HotspotCluster": {
        "properties": {
          "cells": {
            "items": {
              "type": "string"
            },
            "title": "Cells",
            "type": "array"
          },
          "centroidLat": {
            "title": "Centroidlat",
            "type": "number"
          },
          "centroidLon": {
            "title": "Centroidlon",
            "type": "number"
          },
          "clusterId": {
            "title": "Clusterid",
            "type": "integer"
          },
          "meanScore": {
            "title": "Meanscore",
            "type": "number"
          },
          "size": {
            "title": "Size",
            "type": "integer"
          }
        },
        "required": [
          "clusterId",
          "size",
          "meanScore",
          "centroidLat",
          "centroidLon",
          "cells"
        ],
        "title": "HotspotCluster",
        "type": "object"
      },
      "HotspotStatCell": {
        "description": "One classified cell. Only the fields relevant to the active\nmethod are populated: zScore/pValue/confidence for Gi*, clusterId\n(+confidence) for DBSCAN, classification for all methods.",
        "properties": {
          "classification": {
            "title": "Classification",
            "type": "string"
          },
          "clusterId": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Clusterid"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "h3Index": {
            "title": "H3Index",
            "type": "string"
          },
          "pValue": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pvalue"
          },
          "score": {
            "title": "Score",
            "type": "number"
          },
          "zScore": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Zscore"
          }
        },
        "required": [
          "h3Index",
          "score",
          "classification"
        ],
        "title": "HotspotStatCell",
        "type": "object"
      },
      "HotspotsRequest": {
        "properties": {
          "eps_km": {
            "default": 1.5,
            "maximum": 10.0,
            "minimum": 0.1,
            "title": "Eps Km",
            "type": "number"
          },
          "k": {
            "default": 2,
            "maximum": 4.0,
            "minimum": 1.0,
            "title": "K",
            "type": "integer"
          },
          "method": {
            "default": "gi_star",
            "enum": [
              "gi_star",
              "dbscan",
              "binning"
            ],
            "title": "Method",
            "type": "string"
          },
          "min_samples": {
            "default": 4,
            "maximum": 50.0,
            "minimum": 2.0,
            "title": "Min Samples",
            "type": "integer"
          },
          "preset": {
            "default": "warehouse",
            "enum": [
              "warehouse",
              "retail",
              "ev"
            ],
            "title": "Preset",
            "type": "string"
          },
          "threshold": {
            "default": 70.0,
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Threshold",
            "type": "number"
          },
          "weights": {
            "anyOf": [
              {
                "additionalProperties": {
                  "minimum": 0.0,
                  "type": "number"
                },
                "propertyNames": {
                  "enum": [
                    "demographics",
                    "transport",
                    "poi",
                    "zoning",
                    "flood",
                    "aqi"
                  ]
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weights"
          }
        },
        "title": "HotspotsRequest",
        "type": "object"
      },
      "HotspotsResponse": {
        "properties": {
          "cells": {
            "items": {
              "$ref": "#/components/schemas/HotspotStatCell"
            },
            "title": "Cells",
            "type": "array"
          },
          "clusters": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/HotspotCluster"
            },
            "title": "Clusters",
            "type": "array"
          },
          "datasetId": {
            "title": "Datasetid",
            "type": "string"
          },
          "h3Resolution": {
            "title": "H3Resolution",
            "type": "integer"
          },
          "method": {
            "title": "Method",
            "type": "string"
          },
          "preset": {
            "enum": [
              "warehouse",
              "retail",
              "ev"
            ],
            "title": "Preset",
            "type": "string"
          },
          "underserved": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/UnderservedCell"
            },
            "title": "Underserved",
            "type": "array"
          }
        },
        "required": [
          "datasetId",
          "h3Resolution",
          "preset",
          "method",
          "cells"
        ],
        "title": "HotspotsResponse",
        "type": "object"
      },
      "Point": {
        "properties": {
          "lat": {
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Lat",
            "type": "number"
          },
          "lon": {
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Lon",
            "type": "number"
          }
        },
        "required": [
          "lat",
          "lon"
        ],
        "title": "Point",
        "type": "object"
      },
      "ScoreRequest": {
        "properties": {
          "point": {
            "$ref": "#/components/schemas/Point"
          },
          "preset": {
            "default": "warehouse",
            "enum": [
              "warehouse",
              "retail",
              "ev"
            ],
            "title": "Preset",
            "type": "string"
          },
          "weights": {
            "anyOf": [
              {
                "additionalProperties": {
                  "minimum": 0.0,
                  "type": "number"
                },
                "propertyNames": {
                  "enum": [
                    "demographics",
                    "transport",
                    "poi",
                    "zoning",
                    "flood",
                    "aqi"
                  ]
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weights"
          }
        },
        "required": [
          "point"
        ],
        "title": "ScoreRequest",
        "type": "object"
      },
      "ScoreResponse": {
        "properties": {
          "constraints": {
            "items": {
              "$ref": "#/components/schemas/ConstraintResult"
            },
            "title": "Constraints",
            "type": "array"
          },
          "eligible": {
            "title": "Eligible",
            "type": "boolean"
          },
          "h3_index": {
            "title": "H3 Index",
            "type": "string"
          },
          "lat": {
            "title": "Lat",
            "type": "number"
          },
          "lon": {
            "title": "Lon",
            "type": "number"
          },
          "score": {
            "title": "Score",
            "type": "number"
          },
          "subscores": {
            "$ref": "#/components/schemas/Subscores"
          }
        },
        "required": [
          "h3_index",
          "lat",
          "lon",
          "score",
          "eligible",
          "subscores",
          "constraints"
        ],
        "title": "ScoreResponse",
        "type": "object"
      },
      "Subscores": {
        "properties": {
          "aqi": {
            "title": "Aqi",
            "type": "number"
          },
          "demographics": {
            "title": "Demographics",
            "type": "number"
          },
          "flood": {
            "title": "Flood",
            "type": "number"
          },
          "poi": {
            "title": "Poi",
            "type": "number"
          },
          "transport": {
            "title": "Transport",
            "type": "number"
          },
          "zoning": {
            "title": "Zoning",
            "type": "number"
          }
        },
        "required": [
          "demographics",
          "transport",
          "poi",
          "zoning",
          "flood",
          "aqi"
        ],
        "title": "Subscores",
        "type": "object"
      },
      "UnderservedCell": {
        "properties": {
          "demand": {
            "title": "Demand",
            "type": "number"
          },
          "h3Index": {
            "title": "H3Index",
            "type": "string"
          },
          "supply": {
            "title": "Supply",
            "type": "integer"
          }
        },
        "required": [
          "h3Index",
          "demand",
          "supply"
        ],
        "title": "UnderservedCell",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Wherehouse Geo API",
    "version": "0.3.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/health": {
      "get": {
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health"
      }
    },
    "/v1/heatmap": {
      "get": {
        "operationId": "get_heatmap_v1_heatmap_get",
        "parameters": [
          {
            "in": "query",
            "name": "preset",
            "required": false,
            "schema": {
              "default": "warehouse",
              "enum": [
                "warehouse",
                "retail",
                "ev"
              ],
              "title": "Preset",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HeatmapResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Heatmap"
      }
    },
    "/v1/hotspots": {
      "post": {
        "description": "Spec section 6: Gi*, DBSCAN, or H3 binning over preset composites.\n\nSubscores come from the heatmap cache; only the weighted composite\nis recalculated per request.",
        "operationId": "compute_hotspots_v1_hotspots_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HotspotsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HotspotsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Compute Hotspots"
      }
    },
    "/v1/presets": {
      "get": {
        "operationId": "get_presets_v1_presets_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "additionalProperties": {
                      "type": "number"
                    },
                    "type": "object"
                  },
                  "title": "Response Get Presets V1 Presets Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Presets"
      }
    },
    "/v1/score": {
      "post": {
        "operationId": "score_point_v1_score_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScoreRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScoreResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Score Point"
      }
    },
    "/v1/score/batch": {
      "post": {
        "operationId": "score_batch_v1_score_batch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchScoreRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchScoreResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Score Batch"
      }
    }
  }
}